chore: 🤖 rename home feature into index feature
