--- src/features/home/components/HomePage/index.ts
@@ -1,2 +1,0 @@
-export * from './HomePage'
-export * from './HomePage.css'
