+++ src/features/index/components/IndexPage/IndexPage.css.ts
