+++ src/features/index/components/IndexPage/index.ts
@@ -1,0 +1,2 @@
+export * from './IndexPage'
+export * from './IndexPage.css'
